fix: [Upload] ts error
